aider: Set `archive_folder` to the `midjourney/dimjournal` subdirectory in the user's Pictures folder if no `archive_folder` argument is provided.
# Context:
USER: archive_folder is currently {     archive_folder = (
        Path(archive_folder) if archive_folder else Path(Path.cwd(), "mj-archive2")
    )
}
ASSISTANT: Yes, that's correct. The `archive_folder` variable is assigned a value based on the `archive_folder` argument. If the `archive_folder` argument is provided, it is converted to a `Path` object. If it is not provided, a new `Path` object is created with the current working directory as the base and "mj-archive2" as the subdirectory.

Do you want to make any changes to this part of the code?
USER: Can we replace it with a cross-platform solution so that the default archive_folder is always the `midjourney/dimjournal` folder within the user's Pictures folder?
ASSISTANT: Sure, we can use the `os` module to get the user's home directory and then append the Pictures folder and the `midjourney/dimjournal` subdirectory to it. Here's how we can do it:

```python
src/dimjournal/dimjournal.py
<<<<<<< ORIGINAL
    archive_folder = (
        Path(archive_folder) if archive_folder else Path(Path.cwd(), "mj-archive2")
    )
=======
    import os
    archive_folder = (
        Path(archive_folder) if archive_folder else Path(os.path.expanduser("~"), "Pictures", "midjourney", "dimjournal")
    )
>>>>>>> UPDATED
```

This will set the `archive_folder` to the `midjourney/dimjournal` subdirectory in the user's Pictures folder if no `archive_folder` argument is provided.