aider: Refactor code to set the `archive_folder` to the user's "Pictures" folder on macOS and Linux, and "My Pictures" folder on Windows.
# Context:
USER: is this cross-platform? Will it work on macOS and Windows and Linux? I mean the Pictures folder which is used by the system
ASSISTANT: Yes, the solution provided is cross-platform. The `os.path.expanduser("~")` function returns the home directory of the current user on any platform (Windows, macOS, Linux). The "Pictures" folder is a standard subdirectory in the home directory on all these platforms.

However, please note that the folder names are case-sensitive on some systems. For example, on macOS and Linux, the folder is named "Pictures", but on Windows, it's "My Pictures". If you want to handle this case, you can use the `os` module to check the operating system and set the folder name accordingly.

Here's how you can do it:

```python
src/dimjournal/dimjournal.py
<<<<<<< ORIGINAL
    import os
    archive_folder = (
        Path(archive_folder) if archive_folder else Path(os.path.expanduser("~"), "Pictures", "midjourney", "dimjournal")
    )
=======
    import os
    pictures_folder = "My Pictures" if os.name == 'nt' else "Pictures"
    archive_folder = (
        Path(archive_folder) if archive_folder else Path(os.path.expanduser("~"), pictures_folder, "midjourney", "dimjournal")
    )
>>>>>>> UPDATED
```

This will set the `archive_folder` to the `midjourney/dimjournal` subdirectory in the user's Pictures folder (or "My Pictures" on Windows) if no `archive_folder` argument is provided.